style: make the three folders city , directorate and neighborhood one folder called pharmacylocation
