+++ app/Http/Controllers/pharmacylocation/NeighborhoodController.php
@@ -1,29 +1,29 @@
 <?php
 
-namespace App\Http\Controllers\directorate;
-use App\Models\Directorate;
+namespace App\Http\Controllers\pharmacylocation;
+use App\Models\Neighborhood;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
-class DirectorateController extends Controller
+class NeighborhoodController extends Controller
 {
 
     public function index()
     {
-      $directorates = Directorate::get();
-      return response($directorates);
-    // return view('directorates', compact('directorates'));
+      $neighborhoods = Neighborhood::get();
+      return response($neighborhoods);
+      // return view('neighborhoods', compact('neighborhoods'));
     }
 
     public function store(Request $request)
     {
       $validator = Validator::make($request->all(), [
         'name' => 'required|min:2|max:30|string',
-        'city_id' => 'required','city_id' => 'required',
+        'directorate_id' => 'required'
       ]);
-      Directorate::create([
+      Neighborhood::create([
         'name' => $request->input('name'),
-        'city_id' => $request->input('city_id'),
+        'directorate_id' => $request->input('directorate_id'),
       ]);
 
       return response(['added successfully', $validator->errors()]);
@@ -32,28 +32,28 @@ class DirectorateController extends Controller
 
     public function show($id)
     {
-      $directorates = Directorate::where('id', $id)->get();
-      return response($directorates);
+      $neighborhoods = Neighborhood::where('id', $id)->get();
+      return response($neighborhoods);
     }
 
     public function update(Request $request, $id)
     {
       $validator = Validator::make($request->all(), [
         'name' => 'required|min:2|max:30|string',
-        'city_id' => 'required',
+        'directorate_id' => 'required',
       ]);
-      Directorate::where('id', $id)
+      Neighborhood::where('id', $id)
         ->update([
           'name' => $request->input('name'),
-          'city_id' => $request->input('city_id')
+          'directorate_id' => $request->input('directorate_id')
         ]);
-       // return redirect()->back()->with('status', 'edit successfully');
-        return response(['edit successfully', $validator->errors()]);
+      //  return redirect()->back()->with('status', 'edit successfully');
+      return response(['edit successfully', $validator->errors()]);
     }
 
     public function destroy($id)
     {
-      return Directorate::where('id', $id)->delete() ? "deleted" : 'not deleted';
+      return Neighborhood::where('id', $id)->delete() ? "deleted" : 'not deleted';
       //return redirect()->back()->with('status', Ad::where('id', $id)->delete() ? "deleted" : 'not deleted');
     }
 }
